Renames census/ to insturmentation/stats and updates package names accordingly
