--- core/java/com/google/census/CensusContextFactory.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2016, Google Inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.census;
-
-import java.nio.ByteBuffer;
-import javax.annotation.Nullable;
-
-/**
- * Factory class for {@link CensusContext}.
- */
-public abstract class CensusContextFactory {
-  /**
-   * Creates a {@link CensusContext} from the given on-the-wire encoded representation.
-   *
-   * <p>Should be the inverse of {@link CensusContext#serialize()}. The serialized representation
-   * should be based on the {@link CensusContext} protobuf representation.
-   *
-   * @param buffer on-the-wire representation of a {@link CensusContext}
-   * @return a {@link CensusContext} deserialized from {@code buffer}
-   */
-  @Nullable
-  public abstract CensusContext deserialize(ByteBuffer buffer);
-
-  /**
-   * Returns the default {@link CensusContext}.
-   */
-  public abstract CensusContext getDefault();
-}
