--- core/java/com/google/census/Metric.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2016, Google Inc.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *     http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.census;
-
-/**
- * Immutable representation of a Metric.
- */
-public class Metric {
-  public Metric(MetricName name, double value) {
-    this.name = name;
-    this.value = value;
-  }
-
-  /**
-   * Extracts the {@link Metric} name.
-   */
-  public MetricName getName() {
-    return name;
-  }
-
-  /**
-   * Extracts the {@link Metric} value.
-   */
-  public double getValue() {
-    return value;
-  }
-
-  private final MetricName name;
-  private final double value;
-}
