+++ core/java/com/google/instrumentation/stats/Census.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2016, Google Inc.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.instrumentation.stats;
+
+/**
+ * {@link Census}.
+ */
+public final class Census {
+  private static final CensusContextFactory CONTEXT_FACTORY = new Provider<CensusContextFactory>(
+      "com.google.instrumentation.stats.CensusContextFactoryImpl").newInstance();
+
+  /**
+   * Returns the default {@link CensusContextFactory}.
+   */
+  public static CensusContextFactory getCensusContextFactory() {
+    return CONTEXT_FACTORY;
+  }
+
+  // VisibleForTesting
+  Census() {
+    throw new AssertionError();
+  }
+}
