+++ core_native/java/com/google/instrumentation/stats/CensusCurrentContext.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2016, Google Inc.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.instrumentation.stats;
+
+/**
+ * Native implementation of thread-local {@link CensusContext}.
+ */
+class CensusCurrentContext {
+  static final ThreadLocal<CensusContextImpl> contexts = new ThreadLocal<CensusContextImpl>() {
+    @Override
+    protected CensusContextImpl initialValue() {
+      return CensusContextFactoryImpl.DEFAULT;
+    }
+  };
+
+  public static void set(CensusContextImpl context) {
+    contexts.set(context);
+  }
+
+  public static CensusContextImpl get() {
+    return contexts.get();
+  }
+}
